Moved InitPayload from graphql to handler package, updated test to import it from there.
