+++ graphql/context.go
@@ -195,55 +195,3 @@ func (c *RequestContext) RegisterExtension(key string, value interface{}) error 
 	c.Extensions[key] = value
 	return nil
 }
-
-const (
-	initpayload key = "initpayload_context"
-)
-
-// InitPayload is a structure that is parsed from the websocket init message payload. TO use
-// request headers for non-websocket, instead wrap the graphql handler in a middleware.
-type InitPayload map[string]interface{}
-
-// GetString safely gets a string value from the payload. It returns an empty string if the
-// payload is nil or the value isn't set.
-func (payload InitPayload) GetString(key string) string {
-	if payload == nil {
-		return ""
-	}
-
-	if value, ok := payload[key]; ok {
-		res, _ := value.(string)
-		return res
-	}
-
-	return ""
-}
-
-// Authorization is a short hand for getting the Authorization header from the
-// payload.
-func (payload InitPayload) Authorization() string {
-	if value := payload.GetString("Authorization"); value != "" {
-		return value
-	}
-
-	if value := payload.GetString("authorization"); value != "" {
-		return value
-	}
-
-	return ""
-}
-
-// WithInitPayload makes a context with the init payload.
-func WithInitPayload(ctx context.Context, payload InitPayload) context.Context {
-	return context.WithValue(ctx, initpayload, payload)
-}
-
-// GetInitPayload gets the payload from context.
-func GetInitPayload(ctx context.Context) InitPayload {
-	payload, ok := ctx.Value(initpayload).(InitPayload)
-	if !ok {
-		return nil
-	}
-
-	return payload
-}
